+++ kroxylicious-microbenchmarks/src/main/java/io/kroxylicious/benchmarking/jmh/InvokerScalabilityBenchmark.java
@@ -4,7 +4,7 @@
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
 
-package io.kroxylicious.microbenchmarks;
+package io.kroxylicious.benchmarking.jmh;
 
 import java.util.concurrent.TimeUnit;
 
@@ -19,10 +19,8 @@ import org.openjdk.jmh.annotations.Threads;
 import org.openjdk.jmh.annotations.Warmup;
 import org.openjdk.jmh.infra.Blackhole;
 
-import io.kroxylicious.benchmarking.jmh.ArrayFilterInvoker;
-import io.kroxylicious.benchmarking.jmh.SpecificFilterInvoker;
-import io.kroxylicious.microbenchmarks.filters.TwoInterfaceFilter0;
-import io.kroxylicious.microbenchmarks.filters.TwoInterfaceFilter1;
+import io.kroxylicious.benchmarking.jmh.filters.TwoInterfaceFilter0;
+import io.kroxylicious.benchmarking.jmh.filters.TwoInterfaceFilter1;
 import io.kroxylicious.proxy.filter.Filter;
 import io.kroxylicious.proxy.internal.filter.FilterInvoker;
 import io.kroxylicious.proxy.internal.filter.FilterInvokers;
