+++ kroxylicious-microbenchmarks/src/main/java/io/kroxylicious/benchmarking/jmh/filters/OneInterfaceFilter.java
@@ -4,7 +4,7 @@
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
 
-package io.kroxylicious.microbenchmarks.filters;
+package io.kroxylicious.benchmarking.jmh.filters;
 
 import java.util.concurrent.CompletionStage;
